move selectors, add lighthouse
