+++ netlify.toml
@@ -24,3 +24,6 @@
 
 [build.environment]
   NODE_OPTIONS = "--max_old_space_size=4096"
+
+[[plugins]]
+  package = "@netlify/plugin-lighthouse"
