+++ package.json
@@ -126,6 +126,7 @@
   "devDependencies": {
     "@babel/preset-typescript": "7.18.6",
     "@craco/craco": "7.0.0",
+    "@netlify/plugin-lighthouse": "4.0.3",
     "@types/debounce": "1.2.1",
     "@types/jest": "27.5.2",
     "@types/lodash": "4.14.189",
