--- src/components/Form/TextArea/TextArea.styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const Text = styled.div`
-  & textarea {
-    border-radius: 0;
-  }
-`
