+++ src/components/TypeWriter/TypeWriter.styles.ts
