+++ src/components/TypeWriter/TypeWriter.tsx
