+++ src/services/index.ts
@@ -1,2 +1,2 @@
-export { default as AccountUpdateService } from './account.update.service'
-export { default as EntityUpdateService } from './entity.update.service'
+export { default as AccountUpdateService } from './account'
+export { default as EntityUpdateService } from './entity'
